refactoring removing bugs

+++ aula/apps/presencia/business_rules/impartir.py
@@ -5,6 +5,7 @@ from django.apps import apps
 from aula.apps.aules.models import ReservaAula, Aula
 from aula.apps.horaris.models import FranjaHoraria
 from aula.apps.missatgeria.models import Missatge
+from django.contrib.auth.models import User
 
 def impartir_clean( instance ):
     pass
@@ -24,11 +25,15 @@ def impartir_pre_save(sender, instance,  **kwargs):
 
     instance.clean()
 
-    reserva_compartida = (instance.reserva is not None 
+    print instance.pk
+
+    reserva_compartida = (bool(instance.reserva)
                           and instance.reserva.impartir_set.count() > 1 )
     if (not reserva_compartida 
         and instance.reserva is not None 
-        and not instance.es_reserva_manual ):
+        and not instance.reserva.es_reserva_manual ):
+        fake_l4_credentials = (None, True)
+        instance.reserva.credentials = fake_l4_credentials
         instance.reserva.delete()
 
 
@@ -45,15 +50,24 @@ def impartir_post_save(sender, instance, created, **kwargs):
                                                dia_reserva=instance.dia_impartir,
                                                hora=instance.horari.hora )
 
-        reserves_manuals = list( reserves.filter( impartir__isnull = True ) )
+        reserves_manuals = list( reserves.filter( es_reserva_manual = True ) )
         usuari_notificacions, new = User.objects.get_or_create( username = 'TP')
         for reserva in reserves_manuals:
-            reserva.delete()
-            msg = Missatge(
-                remitent=usuari_notificacions,
-                text_missatge=u"El sistema ha hagut d'anul·lar la teva reserva d'aula: {0}".format(reserva),
-                )
-            msg.envia_a_usuari(reserva.usuari,'VI')            
+            impartir = reserva.impartir_set.first()
+            if impartir:
+                msg = u"El sistema ha hagut d'anul·lar aquest reserva: {0}".format(impartir)
+                reserva.aula = impartir.horari.aula()
+                reserva.es_reserva_manual = False
+                fake_l4_credentials = (None, True ) 
+                reserva.credentials = fake_l4_credentials
+                reserva.save()
+                msg = Missatge(
+                    remitent=usuari_notificacions,
+                    text_missatge=msg,
+                    )
+                msg.envia_a_usuari(reserva.usuari,'VI')                   
+            else:
+                reserva.delete()        
 
         reserves_automatiques = list( reserves.filter(impartir__isnull=False) )
         if bool(reserves_automatiques):
